# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/patrols/checkpoints": {
      "get": {
        "operationId": "PatrolsController_getCheckpoints",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get all active patrol checkpoints",
        "tags": [
          "Patrols"
        ]
      }
    },
    "/patrols/visit": {
      "post": {
        "operationId": "PatrolsController_visitCheckpoint",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisitCheckpointDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Visit a checkpoint (Geofence Verified)",
        "tags": [
          "Patrols"
        ]
      }
    },
    "/patrols/history": {
      "get": {
        "operationId": "PatrolsController_getVisitHistory",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get visited checkpoints for current active shift",
        "tags": [
          "Patrols"
        ]
      }
    },
    "/patrols/summary": {
      "get": {
        "operationId": "PatrolsController_getPatrolSummary",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get patrol progress summary for current shift",
        "tags": [
          "Patrols"
        ]
      }
    },
    "/patrols/report": {
      "post": {
        "operationId": "PatrolsController_createPatrolReport",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatrolReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Submit a Patrol Report Finding",
        "tags": [
          "Patrols"
        ]
      },
      "get": {
        "operationId": "PatrolsController_getPatrolReports",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get List of Patrol Reports",
        "tags": [
          "Patrols"
        ]
      }
    },
    "/patrols/report/{patrol_id}": {
      "get": {
        "operationId": "PatrolsController_getPatrolReportDetail",
        "parameters": [
          {
            "name": "patrol_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Patrol Report Detail",
        "tags": [
          "Patrols"
        ]
      }
    },
    "/regu": {
      "post": {
        "operationId": "ReguController_createRegu",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReguDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a new Regu (Koordinator Only)",
        "tags": [
          "Regu (Squads)"
        ]
      },
      "get": {
        "operationId": "ReguController_getRegus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get list of all Regu",
        "tags": [
          "Regu (Squads)"
        ]
      }
    },
    "/regu/{regu_id}": {
      "get": {
        "operationId": "ReguController_getReguById",
        "parameters": [
          {
            "name": "regu_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Regu details and members",
        "tags": [
          "Regu (Squads)"
        ]
      },
      "put": {
        "operationId": "ReguController_updateRegu",
        "parameters": [
          {
            "name": "regu_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReguDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update Regu (Koordinator Only)",
        "tags": [
          "Regu (Squads)"
        ]
      },
      "delete": {
        "operationId": "ReguController_deleteRegu",
        "parameters": [
          {
            "name": "regu_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete Regu (Koordinator Only)",
        "tags": [
          "Regu (Squads)"
        ]
      }
    },
...
    "/performance/evaluations": {
      "get": {
        "operationId": "PerformanceController_getAllEvaluations",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get Performance Evaluations of All Linmas (Koordinator)",
        "tags": [
          "Performance Evaluation"
        ]
      }
    },
    "/performance/my-evaluation": {
      "get": {
        "operationId": "PerformanceController_getMyEvaluation",
        "parameters": [
          {
            "name": "month",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "year",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get My Performance Evaluation (Linmas)",
        "tags": [
          "Performance Evaluation"
        ]
      }
    }
  },
  "info": {
    "title": "MANALINMAS API",
    "description": "Manajemen Kinerja Linmas & Pelaporan Darurat API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "VisitCheckpointDto": {
        "type": "object",
        "properties": {
          "checkpoint_id": {
            "type": "string",
            "example": "uuid-checkpoint-id"
          },
          "latitude": {
            "type": "number",
            "example": -6.90481
          },
          "longitude": {
            "type": "number",
            "example": 107.61031
          }
        },
        "required": [
          "checkpoint_id",
          "latitude",
          "longitude"
        ]
      },
      "CreatePatrolReportDto": {
        "type": "object",
        "properties": {
          "patrol_type": {
            "type": "string",
            "enum": [
              "SUSPICIOUS_OBJECT",
              "FACILITY_DAMAGE",
              "POTENTIAL_HAZARD"
            ],
            "example": "FACILITY_DAMAGE"
          },
          "description": {
            "type": "string",
            "example": "Lampu jalan mati di Blok B3"
          },
          "latitude": {
            "type": "number",
            "example": -6.96845
          },
          "longitude": {
            "type": "number",
            "example": 107.5922
          },
          "photo": {
            "type": "string",
            "format": "binary",
            "description": "Foto temuan"
          }
        },
        "required": [
          "patrol_type",
          "description",
          "latitude",
          "longitude",
          "photo"
        ]
      },
      "CreateReguDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Regu 1"
          },
          "description": {
            "type": "string",
            "example": "  "
          },
          "linmas_ids": {
            "example": [
              "uuid-linmas-1",
              "uuid-linmas-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateReguDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Regu 2 Update"
          },
          "description": {
            "type": "string",
            "example": "update"
          },
          "linmas_ids": {
            "example": [
              "uuid-linmas-1",
              "uuid-linmas-2"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}